+++ src/pydvl/value/shapley/owen.py
@@ -1,21 +1,16 @@
 import operator
 from enum import Enum
 from functools import reduce
+from itertools import cycle, takewhile
 from typing import Sequence
 
 import numpy as np
-from _warnings import warn
 from numpy.typing import NDArray
+from tqdm import tqdm
 
-from pydvl.utils import (
-    MapReduceJob,
-    ParallelConfig,
-    Status,
-    Utility,
-    maybe_progress,
-    random_powerset,
-)
+from pydvl.utils import MapReduceJob, ParallelConfig, Utility, random_powerset
 from pydvl.value import ValuationResult
+from pydvl.value.stopping import MinUpdates
 
 
 class OwenAlgorithm(Enum):
@@ -52,32 +47,37 @@ def _owen_sampling_shapley(
     :param job_id: For positioning of the progress bar
     :return: Object with the data values, errors.
     """
-    values = np.zeros(len(u.data))
-
     q_stop = {OwenAlgorithm.Standard: 1.0, OwenAlgorithm.Antithetic: 0.5}
     q_steps = np.linspace(start=0, stop=q_stop[method], num=max_q)
 
-    index_set = set(indices)
-    for i in maybe_progress(indices, progress, position=job_id):
+    result = ValuationResult.empty(
+        algorithm="owen_sampling_shapley_" + str(method), indices=indices
+    )
+
+    done = MinUpdates(1)
+    repeat_indices = takewhile(lambda _: not done(result), cycle(indices))
+    pbar = tqdm(disable=not progress, position=job_id, total=100, unit="%")
+    for idx in repeat_indices:
+        pbar.n = 100 * done.completion()
+        pbar.refresh()
         e = np.zeros(max_q)
-        subset = np.array(list(index_set.difference({i})))
+        subset = np.setxor1d(u.data.indices, [idx], assume_unique=True)
         for j, q in enumerate(q_steps):
             for s in random_powerset(subset, n_samples=n_samples, q=q):
-                marginal = u({i}.union(s)) - u(s)
+                marginal = u({idx}.union(s)) - u(s)
                 if method == OwenAlgorithm.Antithetic and q != 0.5:
-                    s_complement = index_set.difference(s)
-                    marginal += u({i}.union(s_complement)) - u(s_complement)
+                    s_complement = np.setxor1d(subset, s, assume_unique=True)
+                    marginal += u({idx}.union(s_complement)) - u(s_complement)
+                    marginal /= 2
                 e[j] += marginal
         e /= n_samples
-        # values[i] = e.mean()
+        result.update(idx, e.mean())
         # Trapezoidal rule
-        values[i] = (e[:-1] + e[1:]).sum() / (2 * max_q)
+        # TODO: investigate whether this or other quadrature rules are better
+        #  than a simple average
+        # result.update(idx, (e[:-1] + e[1:]).sum() / (2 * max_q))
 
-    return ValuationResult(
-        algorithm="owen_sampling_shapley_" + str(method),
-        status=Status.Converged,
-        values=values,
-    )
+    return result
 
 
 def owen_sampling_shapley(
@@ -92,9 +92,6 @@ def owen_sampling_shapley(
 ) -> ValuationResult:
     r"""Owen sampling of Shapley values as described in
     :footcite:t:`okhrati_multilinear_2021`.
-
-    .. warning::
-       Antithetic sampling is unstable and not properly tested
 
     This function computes a Monte Carlo approximation to
 
@@ -111,15 +108,18 @@ def owen_sampling_shapley(
     sample $x \in S^{(q_j)}$ if a draw from a $Ber(q_j)$ distribution is 1.
 
     The second method, selected with the argument ``mode =
-    OwenAlgorithm.Anthithetic``,
-    uses correlated samples in the inner sum to reduce the variance:
+    OwenAlgorithm.Antithetic``, uses correlated samples in the inner sum to
+    reduce the variance:
 
-    $$\hat{v}_u(i) = \frac{1}{Q M} \sum_{j=0}^Q \sum_{m=1}^M [u(S^{(q_j)}_m
+    $$\hat{v}_u(i) = \frac{1}{2 Q M} \sum_{j=0}^Q \sum_{m=1}^M [u(S^{(q_j)}_m
     \cup \{i\}) - u(S^{(q_j)}_m) + u((S^{(q_j)}_m)^c \cup \{i\}) - u((S^{(
     q_j)}_m)^c)],$$
 
     where now $q_j = \frac{j}{2Q} \in [0,\frac{1}{2}]$, and $S^c$ is the
     complement of $S$.
+
+    .. note::
+       The outer integration could be done instead with a quadrature rule.
 
     :param u: :class:`~pydvl.utils.utility.Utility` object holding data, model
         and scoring function.
@@ -138,13 +138,10 @@ def owen_sampling_shapley(
 
     .. versionadded:: 0.3.0
 
+    .. versionchanged:: 0.5.0
+       Support for parallel computation and enable antithetic sampling.
+
     """
-    if n_jobs > 1:
-        raise NotImplementedError("Parallel Owen sampling not implemented yet")
-
-    if OwenAlgorithm(method) == OwenAlgorithm.Antithetic:
-        warn("Owen antithetic sampling not tested and probably bogus")
-
     map_reduce_job: MapReduceJob[NDArray, ValuationResult] = MapReduceJob(
         u.data.indices,
         map_func=_owen_sampling_shapley,